+++ Visual Studio Projects/ZWaveJS.NET/Demo Application/Nodes.xaml.cs
@@ -1,0 +1,9 @@
+namespace Demo_Application;
+
+public partial class Nodes : ContentPage
+{
+	public Nodes()
+	{
+		InitializeComponent();
+	}
+}